+++ test/t_connectors/t_user.py
